refactor: 특정 주식 가격 정보 recoil에 대한 selector 생성
